card-payment-node-mongo-app
Create a card management program in angular: As a user, I should be able to add a new credit/debit card, delete it and update only Name and expiry of the card. If card is expired, date should be highlighted in red to indicate that expiry needs to be updated.